Telemetry (optional).
After restarting the unit, ask systemctl if the unit `is-failed`.

If the unit has failed, or the initial `try-restart` fails, we
report the issue through telem-record-gen. If no telemetry is
installed, or telemetry is disabled, this should not cause any
errors or failures, and the program should continue normal
execution, albeit with error messages about the failures.